test(resources): split the build and network tests into their own file
The added assertions took build_resources.rs to 508 code lines, past the 500
hard limit. Second time today I measured that after pushing rather than before,
so this is a rebuke to myself as much as a commit message: the check is one
command and the gate is not a substitute for running it.

The cut is the file's own "Build" section marker. Configs, container options and
the driver_opts volume stay; build, networks, the long-form secret/config refs
and the orphan sweep move to build_images.rs. 154 and 355 code lines.
